Change language in F-Zero AX
If the Language setting is English, the game will be in English. If not, it will be in Japanese.
Also, the way the game will be set to English is now more proper, as it will no longer bruteforce a value into a memory address but it changes the emulated Triforce system's regional setting. As a result, some text legibility in English will be fixed (e.g., the Drift Turn text box in the Attract mode will no longer go out of the screen's bounds).

To avoid weird behaviour when the Language setting is Auto, it's better to implement the "Always default to English if the game is not European" pull request as well.